Define and configure all nRF24 pins
NB: SS pin should be configured as output even if we don't use it. Our AVR MCU is SPI
master, so we can use any available output pin as CSN signal. However SS should be
configured as output anyway in order to avoid floating levels and unexpected SPI behavior.

At the moment IRQ pin is configured as input but not yet used.

Signed-off-by: matsi <matsi@example.org>